chore: improve ColorModeProvider, ControlBox exports
